Moved mock into tests folder, dropped mocks package
