--- src/Database/Mock/composer.json
@@ -1,28 +1,0 @@
-{
-    "name": "tenancy/db-driver-mock",
-    "description": "The tenancy/tenancy database driver for mocking",
-    "keywords": ["database", "tenancy", "mock"],
-    "license": "MIT",
-    "require": {
-        "tenancy/framework": "*"
-    },
-    "autoload": {
-        "psr-4": {
-            "Tenancy\\Database\\Drivers\\Mock\\": ""
-        }
-    },
-    "authors": [
-        {
-            "name": "Daniël Klabbers",
-            "email": "daniel@klabbers.email",
-            "homepage": "http://luceos.com"
-        }
-    ],
-    "extra": {
-        "laravel": {
-            "providers": [
-                "Tenancy\\Database\\Drivers\\Mock\\Providers\\DatabaseProvider"
-            ]
-        }
-    }
-}
